refactor(lights): move logic to a new module
this keep things waaaay cleaner!
